# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lucy API",
    "version": "1.2.0"
  },
  "paths": {
    "/prompt_builder": {
      "post": {
        "summary": "Get the prompt that would be used given an agentName, accountName, selectType and query text",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agentName": { "type": "string" },
                  "accountName": { "type": "string" },
                  "selectType": { "type": "string" },
                  "query": { "type": "string" },
                  "conversationId": { "type": "string" }
                },
                "required": [
                  "agentName",
                  "accountName",
                  "selectType",
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of prompt elements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role": { "type": "string" },
                      "content": { "type": "string" }
                    },
                    "required": ["role", "content"]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing agentName, accountName, selectType or query)"
          }
        }
      }
    },
    "/ask": {
      "post": {
        "summary": "Ask a question",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": { "type": "string" },
                  "agentName": { "type": "string", "description": "Explicit agent, or fallback when routing is auto" },
                  "routing": { "type": "string", "enum": ["explicit", "auto"], "default": "explicit" },
                  "contextName": { "type": "string", "description": "Nonblank explicit context bypasses routing; omit or send blank for automatic selection" },
                  "accountName": { "type": "string" },
                  "conversationId": { "type": "string" }
                },
                "required": ["question", "accountName"],
                "anyOf": [
                  { "required": ["agentName"] },
                  { "required": ["routing"], "properties": { "routing": { "enum": ["auto"] } } }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI assistant's response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "string" },
                    "conversation_id": { "type": "string" },
                    "needs_clarification": { "type": "boolean" },
                    "routing": { "type": "object", "description": "Routing decision and classifier call/latency diagnostics when opted in. reason=direct_answer means the routing LLM answered without invoking the agent prompt builder or tool loop." }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing fields, invalid routing mode, or invalid agent)"
          }
        }
      }
    },
...
    "/chats": {
      "post": {
        "summary": "Create a new chat session",
        "description": "Manual creation requires agentName. With routing auto, an omitted or empty agentName stores the configured default agent (normally lucy).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agentName": { "type": "string", "description": "Required unless routing is auto; must be a real configured agent when supplied." },
                  "routing": { "type": "string", "enum": ["explicit", "auto"], "default": "explicit" },
                  "accountName": { "type": "string" },
                  "friendlyName": { "type": "string" },
                  "tags": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                },
                "required": ["accountName"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created chat session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "account_name": { "type": "string" },
                    "agent_name": { "type": "string" },
                    "friendly_name": { "type": "string" },
                    "created_at": { "type": "string", "format": "date-time" },
                    "updated_at": { "type": "string", "format": "date-time" },
                    "tags": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "summary": { "type": "string" },
                    "importance_score": { "type": "number" },
                    "include_in_context": { "type": "boolean" },
                    "metadata": { "type": "object" },
                    "messages": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  },
                  "required": [
                    "id",
                    "account_name",
                    "agent_name",
                    "created_at",
                    "updated_at"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing agentName or accountName)"
          }
        }
      },
      "get": {
        "summary": "List chat sessions",
        "parameters": [
          {
            "name": "agentName",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 50 }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of chat sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "account_name": { "type": "string" },
                      "agent_name": { "type": "string" },
                      "friendly_name": { "type": "string" },
                      "created_at": { "type": "string", "format": "date-time" },
                      "updated_at": { "type": "string", "format": "date-time" },
                      "tags": {
                        "type": "array",
                        "items": { "type": "string" }
                      },
                      "summary": { "type": "string" },
                      "importance_score": { "type": "number" },
                      "include_in_context": { "type": "boolean" },
                      "metadata": { "type": "object" },
                      "message_count": { "type": "integer" }
                    },
                    "required": [
                      "id",
                      "account_name",
                      "agent_name",
                      "created_at",
                      "updated_at",
                      "message_count"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing accountName)"
          }
        }
      }
    },
...
    "/chats/{session_id}/messages": {
      "post": {
        "summary": "Append a message to a chat session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "accountName",
            "in": "query",
            "required": false,
            "description": "Owning account; PATCH and message POST also accept it in the JSON body.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "role": { "type": "string" },
                  "content": { "type": "string" },
                  "metadata": { "type": "object" }
                },
                "required": ["role", "content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message appended",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (missing role or content)"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
...
  }
}